+++ jsonargparse_tests/test_subclasses.py
@@ -2272,11 +2272,11 @@ def test_subclass_error_indentation_invalid_init_arg(parser):
     expected = textwrap.dedent("""
     Parser key "val":
       Does not validate against any of the Union subtypes
-      Subtypes: [<class 'NoneType'>, <class 'int'>, <class 'dict'>]
+      Subtypes: [<class 'int'>, <class 'dict'>, <class 'NoneType'>]
       Errors:
-        - Expected a <class 'NoneType'>
         - Expected a <class 'int'>
         - Expected a <class 'dict'>
+        - Expected a <class 'NoneType'>
       Given value type: <class 'str'>
       Given value: abc
     """).strip()